Group 1/crypto-lending-optimizer-main/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyarrow>=12.0.0  # For parquet support

# Machine Learning - Tree-based
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0

# Deep Learning
torch>=2.0.0
tensorflow>=2.13.0

# Time series specific
statsmodels>=0.14.0  # For GARCH models
arch>=5.5.0  # ARCH/GARCH models
prophet>=1.1.0

# Hyperparameter optimization
optuna>=3.2.0
scikit-optimize>=0.9.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
